# openapi.json
{"openapi":"3.0.1","info":{"title":"Stargazer API","description":"ë³? ê´?ì¸? ì¡°ê±´ ë¶„ì„ ë°? ê´?ì¸¡ì? ì¶”ì²œ ?„œë¹„ìŠ¤ API ë¬¸ì„œ","contact":{"name":"Stargazer Team","email":"support@stargazer.com"},"license":{"name":"Apache 2.0","url":"https://www.apache.org/licenses/LICENSE-2.0.html"},"version":"v1.0.0"},"servers":[{"url":"http://localhost:8080","description":"ë¡œì»¬ ê°œë°œ ?„œë²?"},{"url":"https://api.byeolbolil.xyz","description":"?”„ë¡œë•?…˜ ?„œë²?"}],"security":[{"bearer-jwt":[]}],"tags":[{"name":"?¸ì¦?","description":"?šŒ?› ?¸ì¦? ê´?? ¨ API (ë¡œê·¸?¸, ?šŒ?›ê°??…, ?† ?° ?¬ë°œê¸‰, ë¡œê·¸?•„?›ƒ)"},{"name":"?šŒ?›","description":"?šŒ?› ? •ë³? ê´?? ¨ API"},{"name":"ë³? ê´?ì¸?","description":"ë³? ê´?ì¸? ì¡°ê±´ ë¶„ì„ ë°? ?˜ˆë³? ê´?? ¨ API"},{"name":"ê´?ì¸¡ì?","description":"ë³? ê´?ì¸¡ì? ì¡°íšŒ ê´?? ¨ API"}],"paths":{"/api/v1/auth/reissue":{"post":{"tags":["?¸ì¦?"],"summary":"?•¡?„¸?Š¤ ?† ?° ?¬ë°œê¸‰","description":"ë¦¬í”„? ˆ?‹œ ?† ?°?„ ?‚¬?š©?•˜?—¬ ?ƒˆë¡œìš´ ?•¡?„¸?Š¤ ?† ?°?„ ë°œê¸‰ë°›ìŠµ?‹ˆ?‹¤. ë¦¬í”„? ˆ?‹œ ?† ?°?? ì¿ í‚¤?—?„œ ??™?œ¼ë¡? ?½?–´?˜µ?‹ˆ?‹¤.","operationId":"reissue","responses":{"200":{"description":"?† ?° ?¬ë°œê¸‰ ?„±ê³?","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthTokenResponse"}}}},"401":{"description":"?¸ì¦? ?‹¤?Œ¨ (ë¦¬í”„? ˆ?‹œ ?† ?°?´ ?œ ?š¨?•˜ì§? ?•Šê±°ë‚˜ ë§Œë£Œ?¨)","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthTokenResponse"}}}}}}},"/api/v1/auth/register":{"post":{"tags":["?¸ì¦?"],"summary":"?šŒ?›ê°??…","description":"?ƒˆë¡œìš´ ?šŒ?›?„ ?“±ë¡í•©?‹ˆ?‹¤. ?„±ê³? ?‹œ ??™?œ¼ë¡? ë¡œê·¸?¸?˜?–´ JWT ?•¡?„¸?Š¤ ?† ?°ê³? ë¦¬í”„? ˆ?‹œ ?† ?°(ì¿ í‚¤)?„ ë°˜í™˜?•©?‹ˆ?‹¤.","operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"200":{"description":"?šŒ?›ê°??… ?„±ê³?","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthTokenResponse"}}}},"400":{"description":"?˜ëª»ëœ ?š”ì²? (?´ë©”ì¼ ?˜•?‹ ?˜¤ë¥?, ?•„?ˆ˜ ?•„?“œ ?ˆ„?½, ?´ë©”ì¼ ì¤‘ë³µ ?“±)","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthTokenResponse"}}}}}}},"/api/v1/auth/logout":{"post":{"tags":["?¸ì¦?"],"summary":"ë¡œê·¸?•„?›ƒ","description":"?˜„?¬ ?‚¬?š©?ë¥? ë¡œê·¸?•„?›ƒ ì²˜ë¦¬?•˜ê³? ë¦¬í”„? ˆ?‹œ ?† ?°?„ ë¬´íš¨?™”?•©?‹ˆ?‹¤. ë¦¬í”„? ˆ?‹œ ?† ?°?? ì¿ í‚¤?—?„œ ??™?œ¼ë¡? ?½?–´?˜µ?‹ˆ?‹¤.","operationId":"logout","responses":{"401":{"description":"?¸ì¦? ?‹¤?Œ¨ (ë¦¬í”„? ˆ?‹œ ?† ?°?´ ?œ ?š¨?•˜ì§? ?•Š?Œ)"},"200":{"description":"ë¡œê·¸?•„?›ƒ ?„±ê³?"}}}},"/api/v1/auth/login":{"post":{"tags":["?¸ì¦?"],"summary":"ë¡œê·¸?¸","description":"?´ë©”ì¼ê³? ë¹„ë?ë²ˆí˜¸ë¡? ë¡œê·¸?¸?•©?‹ˆ?‹¤. ?„±ê³? ?‹œ JWT ?•¡?„¸?Š¤ ?† ?°ê³? ë¦¬í”„? ˆ?‹œ ?† ?°(ì¿ í‚¤)?„ ë°˜í™˜?•©?‹ˆ?‹¤.","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"401":{"description":"?¸ì¦? ?‹¤?Œ¨ (?´ë©”ì¼ ?˜?Š” ë¹„ë?ë²ˆí˜¸ ë¶ˆì¼ì¹?)","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthTokenResponse"}}}},"200":{"description":"ë¡œê·¸?¸ ?„±ê³?","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthTokenResponse"}}}},"400":{"description":"?˜ëª»ëœ ?š”ì²? (?´ë©”ì¼ ?˜•?‹ ?˜¤ë¥?, ?•„?ˆ˜ ?•„?“œ ?ˆ„?½ ?“±)","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthTokenResponse"}}}}}}},"/api/v1/spots":{"get":{"tags":["ê´?ì¸¡ì?"],"summary":"ê´?ì¸¡ì? ì¡°íšŒ","description":"ì§?? •?œ ?œ„ì¹˜ì? ë°˜ê²½ ?‚´?˜ ë³? ê´?ì¸¡ì?ë¥? ì¡°íšŒ?•©?‹ˆ?‹¤. ?¸ì¦? ?—†?´ ?‚¬?š© ê°??Š¥?•©?‹ˆ?‹¤.","operationId":"getObservationSpots","parameters":[{"name":"request","in":"query","description":"ê´?ì¸¡ì? ì¡°íšŒ ?š”ì²? ? •ë³? (?œ„?„, ê²½ë„, ë°˜ê²½)","required":true,"schema":{"$ref":"#/components/schemas/ObservationSpotRequest"}}],"responses":{"200":{"description":"ì¡°íšŒ ?„±ê³?","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ObservationSpotResponse"}}}},"400":{"description":"?˜ëª»ëœ ?š”ì²? (?œ„?„/ê²½ë„ ë²”ìœ„ ì´ˆê³¼, ë°˜ê²½ ë²”ìœ„ ì´ˆê³¼ ?“±)","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ObservationSpotResponse"}}}}}}}},"/api/v1/members/me":{"get":{"tags":["?šŒ?›"],"summary":"?‚´ ? •ë³? ì¡°íšŒ","description":"?˜„?¬ ë¡œê·¸?¸?•œ ?‚¬?š©??˜ ? •ë³´ë?? ì¡°íšŒ?•©?‹ˆ?‹¤. JWT ?† ?° ?¸ì¦ì´ ?•„?š”?•©?‹ˆ?‹¤.","operationId":"getMyInfo","responses":{"401":{"description":"?¸ì¦? ?‹¤?Œ¨ (?† ?°?´ ?œ ?š¨?•˜ì§? ?•Šê±°ë‚˜ ë§Œë£Œ?¨)","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MemberInfoResponse"}}}},"200":{"description":"ì¡°íšŒ ?„±ê³?","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MemberInfoResponse"}}}}},"security":[{"bearer-jwt":[]}]}},"/api/v1/members/exists/email":{"get":{"tags":["?šŒ?›"],"summary":"?´ë©”ì¼ ì¤‘ë³µ ê²?ì¦?","description":"?šŒ?›ê°??… ?‹œ ?´ë©”ì¼ ì¤‘ë³µ ?—¬ë¶?ë¥? ?™•?¸?•©?‹ˆ?‹¤. ?¸ì¦? ?—†?´ ?‚¬?š© ê°??Š¥?•©?‹ˆ?‹¤.","operationId":"validateEmailDuplicated","parameters":[{"name":"email","in":"query","description":"ê²?ì¦í•  ?´ë©”ì¼ ì£¼ì†Œ","required":true,"schema":{"type":"string"},"example":"user@example.com"}],"responses":{"400":{"description":"?˜ëª»ëœ ?š”ì²? (?´ë©”ì¼ ?˜•?‹ ?˜¤ë¥? ?“±)","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MemberValidationResponse"}}}},"200":{"description":"ê²?ì¦? ?™„ë£?","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MemberValidationResponse"}}}}}}},"/api/v1/forecast":{"get":{"tags":["ë³? ê´?ì¸?"],"summary":"ë³? ê´?ì¸? ?˜ˆë³? ì¡°íšŒ","description":"ì§?? •?œ ?œ„ì¹˜ì˜ ?–¥?›„ ë©°ì¹ ê°„ì˜ ë³? ê´?ì¸? ?˜ˆë³´ë?? ì¡°íšŒ?•©?‹ˆ?‹¤. ?‚ ì§œë³„, ?‹œê°„ë?ë³„ë¡œ ?ƒ?„¸?•œ ê´?ì¸? ì¡°ê±´ ? ?ˆ˜?? ? •ë³´ë?? ? œê³µí•©?‹ˆ?‹¤. ?¸ì¦? ?—†?´ ?‚¬?š© ê°??Š¥?•©?‹ˆ?‹¤.","operationId":"getForecast","parameters":[{"name":"request","in":"query","description":"ë³? ê´?ì¸? ?˜ˆë³? ?š”ì²? ? •ë³? (?œ„?„, ê²½ë„, ?‚ ì§?, ?‹œê°? - ?‚ ì§œì? ?‹œê°„ì? ?„ ?ƒ?‚¬?•­)","required":true,"schema":{"$ref":"#/components/schemas/StargazingRequest"}}],"responses":{"400":{"description":"?˜ëª»ëœ ?š”ì²? (?œ„?„/ê²½ë„ ë²”ìœ„ ì´ˆê³¼ ?“±)","content":{"*/*":{"schema":{"$ref":"#/components/schemas/StargazingForecastResponse"}}}},"500":{"description":"?„œë²? ?˜¤ë¥? (?™¸ë¶? API ?˜¸ì¶? ?‹¤?Œ¨ ?“±)","content":{"*/*":{"schema":{"$ref":"#/components/schemas/StargazingForecastResponse"}}}},"200":{"description":"ì¡°íšŒ ?„±ê³?","content":{"*/*":{"schema":{"$ref":"#/components/schemas/StargazingForecastResponse"}}}}}}},"/api/v1/analyze":{"get":{"tags":["ë³? ê´?ì¸?"],"summary":"ë³? ê´?ì¸? ì¡°ê±´ ë¶„ì„","description":"ì§?? •?œ ?œ„ì¹?, ?‚ ì§?, ?‹œê°„ì˜ ë³? ê´?ì¸? ì¡°ê±´?„ ì¢…í•©? ?œ¼ë¡? ë¶„ì„?•©?‹ˆ?‹¤. ê¸°ìƒ ? •ë³?, ì²œë¬¸ ? •ë³?, ê´‘í•´ ? •ë³´ë?? ?¬?•¨?•œ ?ƒ?„¸ ë¶„ì„ ê²°ê³¼ë¥? ? œê³µí•©?‹ˆ?‹¤. ?¸ì¦? ?—†?´ ?‚¬?š© ê°??Š¥?•©?‹ˆ?‹¤.","operationId":"analyzeStargazingCondition","parameters":[{"name":"request","in":"query","description":"ë³? ê´?ì¸? ì¡°ê±´ ë¶„ì„ ?š”ì²? ? •ë³? (?œ„?„, ê²½ë„, ?‚ ì§?, ?‹œê°?)","required":true,"schema":{"$ref":"#/components/schemas/StargazingRequest"}}],"responses":{"200":{"description":"ë¶„ì„ ?„±ê³?","content":{"*/*":{"schema":{"$ref":"#/components/schemas/StargazingAnalyzeResponse"}}}},"500":{"description":"?„œë²? ?˜¤ë¥? (?™¸ë¶? API ?˜¸ì¶? ?‹¤?Œ¨ ?“±)","content":{"*/*":{"schema":{"$ref":"#/components/schemas/StargazingAnalyzeResponse"}}}},"400":{"description":"?˜ëª»ëœ ?š”ì²? (?œ„?„/ê²½ë„ ë²”ìœ„ ì´ˆê³¼, ?‚ ì§?/?‹œê°? ?˜•?‹ ?˜¤ë¥? ?“±)","content":{"*/*":{"schema":{"$ref":"#/components/schemas/StargazingAnalyzeResponse"}}}}}}}},"components":{"schemas":{"AuthTokenResponse":{"type":"object","properties":{"accessToken":{"type":"string","description":"JWT ?•¡?„¸?Š¤ ?† ?°","example":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."}},"description":"?¸ì¦? ?† ?° ?‘?‹µ DTO"},"RegisterRequest":{"required":["email","nickname","password"],"type":"object","properties":{"email":{"type":"string","description":"?´ë©”ì¼ ì£¼ì†Œ","example":"user@example.com"},"password":{"type":"string","description":"ë¹„ë?ë²ˆí˜¸","example":"password123"},"nickname":{"type":"string","description":"?‹‰?„¤?„","example":"ë³„ë³´?Ÿ¬ê°??"}},"description":"?šŒ?›ê°??… ?š”ì²? ? •ë³?"},"LoginRequest":{"required":["email","password"],"type":"object","properties":{"email":{"type":"string","description":"?´ë©”ì¼ ì£¼ì†Œ","example":"user@example.com"},"password":{"type":"string","description":"ë¹„ë?ë²ˆí˜¸","example":"password123"}},"description":"ë¡œê·¸?¸ ?š”ì²? ? •ë³?"},"ObservationSpotRequest":{"required":["lat","lon","radius"],"type":"object","properties":{"lat":{"maximum":90.0,"exclusiveMaximum":false,"minimum":-90.0,"exclusiveMinimum":false,"type":"number","description":"?œ„?„","format":"double","example":37.5665},"lon":{"maximum":180.0,"exclusiveMaximum":false,"minimum":-180.0,"exclusiveMinimum":false,"type":"number","description":"ê²½ë„","format":"double","example":126.978},"radius":{"maximum":500,"exclusiveMaximum":false,"minimum":10,"exclusiveMinimum":false,"type":"integer","description":"ê²??ƒ‰ ë°˜ê²½ (km)","format":"int32","example":50}},"description":"ê´?ì¸¡ì? ì¡°íšŒ ?š”ì²? DTO"},"ObservationSpotResponse":{"type":"object","properties":{"id":{"type":"integer","description":"ê´?ì¸¡ì? ID","format":"int64","example":1},"title":{"type":"string","description":"ê´?ì¸¡ì? ? œëª?","example":"ê°•ì›?„ ?‰ì°½êµ° ??ê´?? ¹"},"address":{"type":"string","description":"ì£¼ì†Œ","example":"ê°•ì›?„ ?‰ì°½êµ° ??ê´?? ¹ë©?"},"latitude":{"type":"number","description":"?œ„?„","format":"double","example":37.5665},"longitude":{"type":"number","description":"ê²½ë„","format":"double","example":126.978},"description":{"type":"string","description":"ê´?ì¸¡ì? ?„¤ëª?","example":"ë³„ì´ ?˜ ë³´ì´?Š” ëª…ì†Œ?…?‹ˆ?‹¤."},"bortleScale":{"type":"integer","description":"ë³´í? ?“±ê¸? (1-9, ?‚®?„?ˆ˜ë¡? ì¢‹ìŒ)","format":"int32","example":3},"isParkingAvailable":{"type":"boolean","description":"ì£¼ì°¨ ê°??Š¥ ?—¬ë¶?","example":true},"isRestroomAvailable":{"type":"boolean","description":"?™”?¥?‹¤ ?´?š© ê°??Š¥ ?—¬ë¶?","example":true},"isCarAccess":{"type":"boolean","description":"ì°¨ëŸ‰ ? ‘ê·? ê°??Š¥ ?—¬ë¶?","example":true},"thumbnailImage":{"type":"string","description":"?¸?„¤?¼ ?´ë¯¸ì? URL","example":"https://example.com/image.jpg"}},"description":"ê´?ì¸¡ì? ? •ë³? ?‘?‹µ DTO"},"MemberInfoResponse":{"type":"object","properties":{"memberId":{"type":"integer","description":"?šŒ?› ID","format":"int64","example":1},"nickname":{"type":"string","description":"?‹‰?„¤?„","example":"ë³„ë³´?Ÿ¬ê°??"}},"description":"?šŒ?› ? •ë³? ?‘?‹µ DTO"},"MemberValidationResponse":{"type":"object","properties":{"validated":{"type":"boolean","description":"ê²?ì¦? ê²°ê³¼ (true: ?‚¬?š© ê°??Š¥, false: ì¤‘ë³µ?¨)","example":true}},"description":"?´ë©”ì¼ ì¤‘ë³µ ê²?ì¦? ?‘?‹µ DTO"},"LocalTime":{"type":"object","properties":{"hour":{"type":"integer","format":"int32"},"minute":{"type":"integer","format":"int32"},"second":{"type":"integer","format":"int32"},"nano":{"type":"integer","format":"int32"}},"description":"ê´?ì¸? ?‹œê°? (HH:mm ?˜•?‹)","example":"22:00"},"StargazingRequest":{"required":["date","lat","lon","time"],"type":"object","properties":{"lat":{"maximum":90.0,"exclusiveMaximum":false,"minimum":-90.0,"exclusiveMinimum":false,"type":"number","description":"?œ„?„","format":"double","example":37.5665},"lon":{"maximum":180.0,"exclusiveMaximum":false,"minimum":-180.0,"exclusiveMinimum":false,"type":"number","description":"ê²½ë„","format":"double","example":126.978},"date":{"type":"string","description":"ê´?ì¸? ?‚ ì§? (yyyy-MM-dd ?˜•?‹)","format":"date","example":"2025-12-25"},"time":{"$ref":"#/components/schemas/LocalTime"}},"description":"ë³? ê´?ì¸? ì¡°ê±´ ë¶„ì„ ?š”ì²? DTO"},"DailyForecast":{"type":"object","properties":{"date":{"type":"string","description":"?‚ ì§? (?š”?¼ ?¬?•¨)","example":"2025-05-20 (ê¸?)"},"sunrise":{"type":"string","description":"?¼ì¶? ?‹œê°?","example":"07:00"},"sunset":{"type":"string","description":"?¼ëª? ?‹œê°?","example":"18:00"},"moonrise":{"type":"string","description":"?›”ì¶? ?‹œê°?","example":"08:00"},"moonset":{"type":"string","description":"?›”ëª? ?‹œê°?","example":"20:00"},"hourlyForecasts":{"type":"array","description":"?‹œê°„ë?ë³? ?˜ˆë³? ëª©ë¡","items":{"$ref":"#/components/schemas/HourlyForecast"}}},"description":"?¼ë³? ?˜ˆë³?"},"HourlyForecast":{"type":"object","properties":{"time":{"type":"string","description":"?‹œê°?","example":"21:00"},"score":{"type":"integer","description":"ê´?ì¸? ? ?ˆ˜ (0-100)","format":"int32","example":85},"reasons":{"type":"array","description":"ê°ì  ?‚¬?œ  ëª©ë¡","example":["êµ¬ë¦„?Ÿ‰?´ ë§ìŒ"],"items":{"type":"string","description":"ê°ì  ?‚¬?œ  ëª©ë¡","example":"[\"êµ¬ë¦„?Ÿ‰?´ ë§ìŒ\"]"}},"starGrade":{"type":"string","description":"ë³? ?“±ê¸?","example":"4.5?“±ê¸?"},"cloudCover":{"type":"integer","description":"êµ¬ë¦„?Ÿ‰ (%)","format":"int32","example":20},"moonPhase":{"type":"string","description":"?‹¬ ?œ„?ƒ","example":"ì´ˆìŠ¹?‹¬"}},"description":"?‹œê°„ë?ë³? ?˜ˆë³?"},"StargazingForecastResponse":{"type":"object","properties":{"dailyForecasts":{"type":"array","description":"?‚ ì§œë³„ ?˜ˆë³? ëª©ë¡","items":{"$ref":"#/components/schemas/DailyForecast"}}},"description":"ë³? ê´?ì¸? ?˜ˆë³? ?‘?‹µ DTO"},"AstronomyInfo":{"type":"object","properties":{"moonPhase":{"type":"string","description":"?‹¬ ?œ„?ƒ","example":"ì´ˆìŠ¹?‹¬"},"sunrise":{"type":"string","description":"?¼ì¶? ?‹œê°?","example":"07:00"},"sunset":{"type":"string","description":"?¼ëª? ?‹œê°?","example":"18:00"},"moonrise":{"type":"string","description":"?›”ì¶? ?‹œê°?","example":"08:00"},"moonset":{"type":"string","description":"?›”ëª? ?‹œê°?","example":"20:00"}},"description":"ì²œë¬¸ ? •ë³?"},"LightPollutionInfo":{"type":"object","properties":{"bortleClass":{"type":"string","description":"ë³´í? ?“±ê¸? ?´?˜?Š¤","example":"Class 3"},"brightness":{"type":"string","description":"ë°ê¸° ?ˆ˜ì¤?","example":"?‚®?Œ"},"limitingMag":{"type":"string","description":"?•œê³? ?“±ê¸?","example":"6?“±ê¸?"}},"description":"ê´‘í•´ ? •ë³?"},"StargazingAnalyzeResponse":{"type":"object","properties":{"date":{"type":"string","description":"ê´?ì¸? ?‚ ì§?","example":"2025-12-19"},"time":{"type":"string","description":"ê´?ì¸? ?‹œê°?","example":"22:00"},"totalScore":{"type":"integer","description":"ì¢…í•© ? ?ˆ˜ (0-100)","format":"int32","example":85},"reasons":{"type":"array","description":"? ?ˆ˜ ê°ì  ?‚¬?œ  ëª©ë¡","example":["êµ¬ë¦„?Ÿ‰?´ ë§ìŒ","?Šµ?„ê°? ?†’?Œ"],"items":{"type":"string","description":"? ?ˆ˜ ê°ì  ?‚¬?œ  ëª©ë¡","example":"[\"êµ¬ë¦„?Ÿ‰?´ ë§ìŒ\",\"?Šµ?„ê°? ?†’?Œ\"]"}},"aiComment":{"type":"string","description":"AI ?•œì¤? ?‰","example":"?˜¤?Š˜ ë°? ë³? ê´?ì¸¡í•˜ê¸? ì¢‹ì? ?‚ ?”¨?…?‹ˆ?‹¤!"},"weather":{"$ref":"#/components/schemas/WeatherInfo"},"astronomy":{"$ref":"#/components/schemas/AstronomyInfo"},"lightPollution":{"$ref":"#/components/schemas/LightPollutionInfo"}},"description":"ë³? ê´?ì¸? ì¡°ê±´ ë¶„ì„ ?‘?‹µ DTO"},"WeatherInfo":{"type":"object","properties":{"cloudIndex":{"type":"integer","description":"êµ¬ë¦„?Ÿ‰ (%)","format":"int32","example":20},"humidityIndex":{"type":"integer","description":"?Šµ?„ (%)","format":"int32","example":60},"visibilityText":{"type":"string","description":"ê°??‹œ?„ ?“±ê¸?","example":"ë§¤ìš° ì¢‹ìŒ"}},"description":"ê¸°ìƒ ? •ë³?"}},"securitySchemes":{"bearer-jwt":{"type":"http","name":"Authorization","in":"header","scheme":"bearer","bearerFormat":"JWT"}}}}