# src/main/java/xyz/ncookie/stargazer/domain/member/dto/response/AuthTokenResponse.java
package xyz.ncookie.stargazer.domain.member.dto.response;

import io.swagger.v3.oas.annotations.media.Schema;

@Schema(description = "인증 토큰 응답 DTO")
public record AuthTokenResponse(
	@Schema(description = "JWT 액세스 토큰", example = "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...")
	String accessToken
) {}